Restructure dumping to append data as well
When I first made the dumping function, it would just open a file based on the chunk name and dump the data.
This doesn't account for existing files, existing chunks within files, etc.
this way when I dump data, it go many ways: creating a file or opening an existing one, appending to an existing chunk or creating a new one, etc.